Feat : initboradDate
