+++ src/main/java/com/sparta/hirello/primary/board/entity/Board.java
@@ -1,0 +1,40 @@
+package com.sparta.hirello.primary.board.entity;
+
+import com.sparta.hirello.primary.column.entity.Columns;
+import com.sparta.hirello.primary.user.entity.User;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+import java.util.List;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor
+@Table(name = "boards")
+public class Board {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long boardId;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    @Column(nullable = false)
+    private String boardName;
+
+    private String headline;
+
+    @OneToMany(mappedBy = "board")
+    private List<Columns> columnsList;
+}
