+++ src/main/java/com/sparta/hirello/primary/board/jpa/BoardRepository.java
@@ -1,0 +1,8 @@
+package com.sparta.hirello.primary.board.jpa;
+
+import com.sparta.hirello.primary.board.entity.Board;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BoardRepository extends JpaRepository<Board, String>, BoardRepositoryCustom {
+
+}
