+++ src/main/java/com/sparta/hirello/primary/board/jpa/BoardRepositoryCustom.java
@@ -1,0 +1,5 @@
+package com.sparta.hirello.primary.board.jpa;
+
+public interface BoardRepositoryCustom {
+
+}
