+++ src/main/java/com/sparta/hirello/primary/board/service/BoardService.java
@@ -1,0 +1,8 @@
+package com.sparta.hirello.primary.board.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class BoardService {
+
+}
