+++ src/main/java/com/sparta/hirello/primary/card/entity/Card.java
@@ -1,20 +1,29 @@
-package com.sparta.hirello.domain.card.entity;
+package com.sparta.hirello.primary.card.entity;
 
-import com.sparta.hirello.domain.comment.entity.Comment;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import com.sparta.hirello.domain.column.entity.Columns;
-import com.sparta.hirello.domain.user.entity.User;
-
+import com.sparta.hirello.primary.column.entity.Columns;
+import com.sparta.hirello.primary.comment.entity.Comment;
+import com.sparta.hirello.primary.user.entity.User;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
 import java.time.LocalDateTime;
 import java.util.List;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
 @NoArgsConstructor
 @Table(name = "cards")
 public class Card {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long cardId;
