+++ src/main/java/com/sparta/hirello/primary/column/entity/Columns.java
@@ -1,0 +1,43 @@
+package com.sparta.hirello.primary.column.entity;
+
+import com.sparta.hirello.primary.board.entity.Board;
+import com.sparta.hirello.primary.card.entity.Card;
+import com.sparta.hirello.primary.user.entity.User;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+import java.util.List;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor
+@Table(name = "columns")
+public class Columns {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long columnId;
+
+    @Column(nullable = false)
+    private String status;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "board_id")
+    private Board board;
+
+    @OneToMany(mappedBy = "columns")
+    private List<Card> cardList;
+}
