+++ src/main/java/com/sparta/hirello/primary/comment/entity/Comment.java
@@ -1,15 +1,25 @@
-package com.sparta.hirello.domain.comment.entity;
+package com.sparta.hirello.primary.comment.entity;
 
-import com.sparta.hirello.domain.card.entity.Card;
-import com.sparta.hirello.domain.column.entity.Columns;
-import com.sparta.hirello.domain.user.entity.User;
-import jakarta.persistence.*;
+import com.sparta.hirello.primary.card.entity.Card;
+import com.sparta.hirello.primary.column.entity.Columns;
+import com.sparta.hirello.primary.user.entity.User;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import jakarta.persistence.Temporal;
+import jakarta.persistence.TemporalType;
+import java.time.LocalDateTime;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
 
 @Entity
 @NoArgsConstructor
@@ -17,6 +27,7 @@ import java.time.LocalDateTime;
 @Table(name = "comments")
 @EntityListeners(AuditingEntityListener.class)
 public class Comment {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long commentId;
